Initial Commit
Similar to the new weekly workout based on Symfony. This is the orginal
application build by Alvin Ng or David Conner at NC State back in 2010?

Instead of completely re-building it, I'm trying to remove the
duplicationt that has been started, and allow the logging in through
Shibboleth as needed.

This initial commit may have files included that don't really need to
be., the complete Zend Framework Folder is omitted.