Organized application, added 3 new fields to user table, cleaned user form values ...
Many, many needed updates to this old Weekly Workout application,
including:
1.) Fixed the variables from not populating, especially the heart rate
array when first logging into Weekly Workout, these view variables were
never set in the original copy and were not working when re-organizating
this application.
2.) Added three additional fields to the users table, a created_at
timestamp, a semester number and semester string to the user table and
list of users that are in the system.
3.) Cleaned up the instructions a little bit under the current users and
administrators listing to show the new semseter field.
4.) Added a proper notification message to the end user when a workout
is properly added in the system, in the past, it would add a workout but
not relay to the user that it was successful.
5.) Fixed start date when looking for getting all workout dates by a
UnityID and start date and end date.
6.) Have a seperate action to handle shibboleth logins that will create
a user account based on temporary information set inside an array. This
information in future commit will pull information from the server to
use Shibboleth Account Information as needed in the creation of the new
user.
7.) Removed the shibboleth controller as everything will be handled
within the LoginController, specifically the shibbolethAction method
that is pulled when login/shibboleth is called.
8.) Removed the check of the Session namespace when using the
WorkoutController @ init as this application has been changed so that it
is not necessarily looking within the session for this information,
keeping this in continues to provide a session not readable error in
Zend.
9.)The boostrap file is commented out on the _initPlugins to allow for
proper changes, will eventually turn this back on, but not entirely sure
how the ACL plugin is even being utilized with this small application.